Add environment passing to zellij subcommand
Add functionality for passing the environment of the shell that
zellij-run is run from to the zellij run subcommand.

The environment variables are passed through `/dev/shm` if available,
with fallback to `/tmp` and final fallback to the current directory.
Access to the environment file is limited to read and write for the user
only (0600).
The environment file is deleted upon completion of the script.

To ensure that the environment file isn't deleted before it's accessed, a flag file
is created and the zellij run subcommand wrapped in a small script for
touching the flag file, signalling completed load of the environment
file.
The flag file is deleted upon completion of the script.